Fix register size
CNTPCT_EL0 is 64 bits CNTFRQ_EL0 is 64 bits as well but bits 32 to 63 are reserved so the freq is stored in the 32 first bits

+++ include/mm.h
@@ -39,8 +39,8 @@ static inline uint32_t mm_freq() {
     return cntfrq_el0;
 }
 
-static inline uint32_t mm_ticks() {
-    uint32_t cntpct_el0;
+static inline uint64_t mm_ticks() {
+    uint64_t cntpct_el0;
 
     asm volatile("mrs %0, cntpct_el0\n" : "=r" (cntpct_el0) : : "memory");
     return cntpct_el0;
